refactor: cleanup round2
