+++ tests/tests_unit/rules/test_models/test_information_rules.py
@@ -5,7 +5,7 @@ import pandas as pd
 import pytest
 from cognite.client import data_modeling as dm
 
-from cognite.neat.rules.models.data_types import String
+from cognite.neat.rules.models.data_types import DataType, String
 from cognite.neat.rules.models.rules import DMSRules
 from cognite.neat.rules.models.rules._base import SheetList
 from cognite.neat.rules.models.rules._information_rules import (
@@ -13,7 +13,6 @@ from cognite.neat.rules.models.rules._information_rules import (
     InformationRules,
     _InformationRulesConverter,
 )
-from cognite.neat.rules.models.rules._types import XSDValueType
 from cognite.neat.utils.spreadsheet import read_individual_sheet
 from tests.config import DOC_RULES
 
@@ -208,7 +207,7 @@ class TestInformationRules:
         assert errors[0]["msg"] == expected_exception
 
     @pytest.mark.parametrize("rules, expected_exception", list(case_insensitive_value_types()))
-    def test_case_insensitivity(self, rules: dict[str, dict[str, Any]], expected_exception: XSDValueType) -> None:
+    def test_case_insensitivity(self, rules: dict[str, dict[str, Any]], expected_exception: DataType) -> None:
         assert InformationRules.model_validate(rules).properties.data[0].value_type == expected_exception
 
     def test_david_as_dms(self, david_spreadsheet: dict[str, dict[str, Any]]) -> None:
